+++ src/plugin/preset_server/preset_server.hpp
@@ -5,7 +5,7 @@
 
 #include "utils/error_notifications.hpp"
 
-#include "common_infrastructure/preset_pack_info.hpp"
+#include "common_infrastructure/preset_bank_info.hpp"
 #include "common_infrastructure/state/state_coding.hpp"
 #include "common_infrastructure/state/state_snapshot.hpp"
 
@@ -38,7 +38,7 @@ struct PresetFolder {
     Set<String> used_tags {};
     Set<String> used_library_authors {};
 
-    Optional<PresetPackInfo> preset_pack_info {}; // From metadata file (primary importance)
+    Optional<PresetBank> preset_bank_info {}; // From metadata file (primary importance)
 
     u64 all_presets_hash {}; // Hash of all presets in this folder.
 
@@ -82,7 +82,7 @@ struct PresetServer {
     ArenaAllocator folder_node_arena {(Allocator&)arena};
     Span<FolderNode> folder_nodes {};
     Span<usize> folder_node_order_indices {};
-    Span<usize> folder_node_preset_pack_indices {};
+    Span<usize> folder_node_preset_bank_indices {};
 
     DynamicSet<u64, NoHash> preset_file_hashes {arena};
     Bitset<ToInt(PresetFormat::Count)> has_preset_type {};
@@ -103,32 +103,32 @@ void ShutdownPresetServer(PresetServer& server);
 void SetExtraScanFolders(PresetServer& server, Span<String const> folders);
 
 // A 'listing' augments a FolderNode with more preset-specific information such as the PresetFolder
-// (immutable), and preset pack info. For convenience, the node's user_data is also points to this listing so
+// (immutable), and preset bank info. For convenience, the node's user_data is also points to this listing so
 // that given only a node you can get the listing.
 struct PresetFolderListing {
     // May be null if the folder does not contain any presets, in which case it's just a node in the folder
     // tree.
     PresetFolder const* folder;
 
-    // This is of secondary importance if preset_pack_info is specified in PresetFolder.
-    PresetPackInfo const* fallback_preset_pack_info;
+    // This is of secondary importance if preset_bank_info is specified in PresetFolder.
+    PresetBank const* fallback_preset_bank_info;
 
     // The node's user_data is this listing.
     FolderNode const& node;
 };
 
 Optional<sample_lib::LibraryIdRef> AllPresetsSingleLibrary(FolderNode const& node);
-PresetPackInfo const* PresetPackInfoForNode(FolderNode const& node);
-PresetPackInfo const* ContainingPresetPackInfo(FolderNode const* node);
+PresetBank const* PresetBankForNode(FolderNode const& node);
+PresetBank const* ContainingPresetBank(FolderNode const* node);
 bool IsInsideFolder(PresetFolderListing const* node, usize folder_node_hash);
 
 struct PresetsSnapshot {
     // Folders that contain presets, sorted. e.i. these will have PresetFolderListing::folder != null.
     Span<PresetFolderListing const*> folders;
 
-    // Root nodes of all preset packs. All presets are guaranteed to be inside one of these nodes. Presets
+    // Root nodes of all preset banks. All presets are guaranteed to be inside one of these nodes. Presets
     // that aren't explicitly put into packs will be smartly grouped into "Miscellaneous Presets" packs.
-    Span<FolderNode const*> preset_packs;
+    Span<FolderNode const*> preset_banks;
 
     // Additional convenience data
     Set<String> used_tags;